Refactor Incident
Moved to common.
