+++ src/Trade360SDK.CustomersApi/Entities/MetadataApi/Responses/GetIncidentsResponse.cs
@@ -1,19 +1,9 @@
-using System;
 using System.Collections.Generic;
+using Trade360SDK.Common.Entities.Incidents;
 
 namespace Trade360SDK.CustomersApi.Entities.MetadataApi.Responses
 {
     public class GetIncidentsResponse {
         public IEnumerable<Incident>? Data { get; set; }
-    }
-        
-    public class Incident
-    {
-        public int SportId { get; set; }
-        public string SportName { get; set; }
-        public int IncidentId { get; set; }
-        public string IncidentName { get; set; }
-        public DateTime LastUpdate { get; set; }
-        public DateTime CreationDate { get; set; }
     }
 }
